Large improvement pack (#192)
* Added DNS server with TCP support
* Added TCP client limit option, also can be used to disable TCP DNS server in case of issues
* Readme and help pages updated
* Combined API with UDP DNS server
* Use minimally necessary public header
* Added TCP only statistic counters
* Added TCP functional tests
* Basic querries with valgrind
  - fragmented TCP request processing
  - testing 4k large response support
* Started to use stricter compiler warnings
* UDP truncation feature
* Enable more compiler warnings
* Replace c-ares deprecated functions
* Reset HTTPS client on timeouts
* systemd service improvements
* Version bump

+++ src/logging.c
@@ -121,18 +121,18 @@ void _log(const char *file, int line, int severity, const char *fmt, ...) {
   if (chars < 0 || chars >= LOG_LINE_SIZE/2) {
     abort();  // must be impossible
   }
-  buff_pos += chars;
+  buff_pos += (uint32_t)chars;
 
   va_list args;
   va_start(args, fmt);
-  // NOLINTNEXTLINE(clang-analyzer-security.insecureAPI.DeprecatedOrUnsafeBufferHandling)
+  // NOLINTNEXTLINE(clang-diagnostic-format-nonliteral,clang-analyzer-security.insecureAPI.DeprecatedOrUnsafeBufferHandling)
   chars = vsnprintf(buff + buff_pos, LOG_LINE_SIZE - buff_pos, fmt, args);
   va_end(args);
 
   if (chars < 0) {
     abort();  // must be impossible
   }
-  buff_pos += chars;
+  buff_pos += (uint32_t)chars;
   if (buff_pos >= LOG_LINE_SIZE) {
     buff_pos = LOG_LINE_SIZE - 1;
     buff[buff_pos - 1] = '$'; // indicate truncation
